% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNalign.R
\name{CCFalign}
\alias{CCFalign}
\title{CCFalign}
\usage{
CCFalign(
  dat,
  min_purity = 0.05,
  max_purity = 0.95,
  max_tcn = 6,
  min_ccf_is_clonal = 0.9,
  min_frac_clonal_is_truncal = 1,
  gurobi_license = ""
)
}
\description{
CCFalign
}
