% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr_lengths_b37}
\alias{chr_lengths_b37}
\title{chr_lengths_b37}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 25 rows and 2 columns.
}
\usage{
data(chr_lengths_b37)
}
\description{
chr_lengths_b37
}
\keyword{datasets}
