% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytoBand_b37}
\alias{cytoBand_b37}
\title{cytoBand_b37}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 862 rows and 5 columns.
}
\usage{
data(cytoBand_b37)
}
\description{
cytoBand_b37
}
\keyword{datasets}
