% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytoBand_hg38}
\alias{cytoBand_hg38}
\title{cytoBand_hg38}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1549 rows and 5 columns.
}
\usage{
data(cytoBand_hg38)
}
\description{
cytoBand_hg38
}
\keyword{datasets}
