% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_with_confidence_intervals.R
\name{estimate_ccf}
\alias{estimate_ccf}
\title{estimate_ccf}
\usage{
estimate_ccf(
  purity,
  t_tcn,
  t_mcn,
  t_alt_count,
  t_depth,
  n_tcn,
  resolution = 0.001
)
}
\description{
Estimate most likely CCF given observed VAF, purity and local ploidy. Based on PMID 25877892
}
