% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicc.R
\name{extract_medicc2_support_tree}
\alias{extract_medicc2_support_tree}
\title{extract_medicc2_support_tree}
\usage{
extract_medicc2_support_tree(tree_xml_file, normal_sample)
}
\description{
extract_medicc2_support_tree
}
