% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{get_ascn_segments}
\alias{get_ascn_segments}
\title{get_ascn_segments}
\usage{
get_ascn_segments(obj, fit_file, normal_sample, min_purity = 0.1)
}
\description{
Extract a table of copy number segments with purity/ploidy-corrected allele-specific copy number
}
