% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CNalign_obj_for_snp_data.R
\name{get_CNalign_obj_for_snp_data}
\alias{get_CNalign_obj_for_snp_data}
\title{get_CNalign_obj_for_snp_data}
\usage{
get_CNalign_obj_for_snp_data(
  sample_map,
  patient,
  sex,
  build,
  ascat_dir = ".",
  data_dir = ".",
  GCcontentfile = NULL,
  replictimingfile = NULL,
  multipcf_penalty = 70,
  multipcf_refine = F,
  multipcf_selectAlg = "exact",
  cleanup = T,
  seed = NA
)
}
\description{
Load and preprocess standard SNP data (whole-exome, standard-depth whole genome). After running ASCAT on each tumor sample run individually (against the single normal sample), this script will load the resulting data for each sample and merge them, and run multi-sample segmentation via ASCAT's multipcf algorithm. It will then format the output for use with CNalign.
}
