% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_fit}
\alias{get_fit}
\title{get_fit}
\usage{
get_fit(
  sample_name,
  obj,
  dipLogR = NA,
  purity = NA,
  ploidy = NA,
  homdel_mb_max = 100,
  neg_mb_max = 0,
  min_ploidy = 1.7,
  max_ploidy = 5,
  min_purity = 0.05,
  max_purity = 0.95,
  cores = 1,
  best_only = T,
  bin_level = F,
  purity_stepsize = 0.025,
  ploidy_stepsize = 0.025
)
}
\description{
get_fit
}
