% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{get_highlighted_segments_for_solution}
\alias{get_highlighted_segments_for_solution}
\title{get_highlighted_segments_for_solution}
\usage{
get_highlighted_segments_for_solution(model, sol_number)
}
\description{
get_highlighted_segments_for_solution
}
