% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicc.R
\name{medicc2_tree_and_heatmap}
\alias{medicc2_tree_and_heatmap}
\title{medicc2_tree_and_heatmap}
\usage{
medicc2_tree_and_heatmap(
  tree_xml_file,
  cn_profile_file,
  groups = NULL,
  title = NA,
  bp_to_mb = F,
  vertical = F,
  ybreaksize = 2,
  build = "hg19"
)
}
\description{
medicc2_tree_and_heatmap
}
