% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_copynumber_angular_distance}
\alias{plot_copynumber_angular_distance}
\title{plot_copynumber_angular_distance}
\usage{
plot_copynumber_angular_distance(
  obj = NULL,
  segs = NULL,
  build = NULL,
  groups = NULL,
  group_colors = NULL,
  normal_name,
  patient = ""
)
}
\description{
Plot a preliminary copy number heatmap based on angular distances from total copy number
}
