% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{plot_fit}
\alias{plot_fit}
\title{plot_fit}
\usage{
plot_fit(
  sample_name,
  fit,
  obj,
  build,
  int_copies = F,
  highlight_seg = c(),
  LogR_min = NA,
  LogR_max = NA,
  point_size = 0.5
)
}
\description{
plot_fit
}
