% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R, R/plot_tcn_angular_distance_heatmap.R
\name{plot_tcn_angular_distance_heatmap}
\alias{plot_tcn_angular_distance_heatmap}
\title{plot_tcn_angular_distance_heatmap}
\usage{
plot_tcn_angular_distance_heatmap(
  obj = NULL,
  segs = NULL,
  build = NULL,
  sample_groups = NULL,
  group_colors = NULL,
  normal_sample,
  patient = ""
)

plot_tcn_angular_distance_heatmap(
  obj = NULL,
  segs = NULL,
  build = NULL,
  sample_groups = NULL,
  group_colors = NULL,
  normal_sample,
  patient = ""
)
}
\description{
Plot a preliminary copy number heatmap based on angular distances from total copy number

Plot a preliminary copy number heatmap based on angular distances from total copy number
}
