% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_for_multipcf.R
\name{prep_data_for_multipcf}
\alias{prep_data_for_multipcf}
\title{prep_data_for_multipcf}
\usage{
prep_data_for_multipcf(
  Tumor_LogR_file,
  Tumor_BAF_file,
  Germline_LogR_file,
  Germline_BAF_file,
  sex,
  build,
  GCcontentfile = NULL,
  replictimingfile = NULL
)
}
\description{
Run ascat.loadData and format a data-object for use with ascat.asmultipcf
}
