% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNAlignR.R
\name{run_CNAlign}
\alias{run_CNAlign}
\title{run_CNAlign}
\usage{
run_CNAlign(
  dat,
  gurobi_license,
  py_script = NA,
  min_ploidy = 1.7,
  max_ploidy = 6,
  min_purity = 0.05,
  max_purity = 0.95,
  min_aligned_seg_mb = 5,
  max_homdel_mb = 100,
  delta_tcn_to_int = 0.2,
  delta_tcn_to_avg = 0.1,
  delta_tcnavg_to_int = 0.1,
  delta_mcn_to_int = 0.2,
  delta_mcn_to_avg = 0.1,
  delta_mcnavg_to_int = 0.1,
  rho = 1,
  timeout = 3 * 60,
  min_cna_segments_per_sample = 1,
  mcn_weight = 0.5,
  obj2_clonalonly = F,
  sol_count = 10,
  max_tcn_avg_int = 100,
  tcn_only = F,
  n_unique_tcn_in_sol = 1
)
}
\description{
Determine purity and ploidy values for multiple tumor samples with shared ancestry. This function uses the GuRoBi solver to determine purity/ploidy values for each sample that will *maximize* the number of segments with the same (allele-specific) integer copy numbers in at least rho% of samples.
}
