% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{write_refphase_segs}
\alias{write_refphase_segs}
\title{write_refphase_segs}
\usage{
write_refphase_segs(segs, cn_events = NULL, file, output_format = "summary")
}
\description{
Output refphase segments as a tsv-file
}
