% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_with_confidence_intervals.R
\name{expected_mutant_copies}
\alias{expected_mutant_copies}
\title{expected_mutant_copies}
\usage{
expected_mutant_copies(t_var_freq, total_copies, purity)
}
\description{
Estimate mutant copy number, given observed VAF, purity, and local ploidy. Based on PMID 28270531
}
