% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CNalign_obj_for_lowpass_data.R
\name{get_CNalign_obj_for_lowpass_data}
\alias{get_CNalign_obj_for_lowpass_data}
\title{get_CNalign_obj_for_lowpass_data}
\usage{
get_CNalign_obj_for_lowpass_data(
  qdnaseq_data,
  pileup_data,
  phased_bcf,
  sample_map,
  patient,
  sex,
  normal_sample,
  build,
  data_dir = ".",
  max_phaseable_distance = 20000,
  min_bin_reads_for_baf = 10,
  blacklisted_regions_file = NA,
  LogR_range_allowed = c(-3, 3),
  LogR_winsor_percentiles = c(NA, NA),
  LogR_smooth_bins = NA,
  multipcf_penalty = 70,
  multipcf_refine = F,
  multipcf_selectAlg = "exact",
  cleanup = T,
  seed = NA
)
}
\description{
Load and preprocess low-pass WGS data from GLIMPSE2, QDNAseq and snp-pileup workflows.
}
