% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_tcn_heatmap}
\alias{plot_tcn_heatmap}
\title{plot_tcn_heatmap}
\usage{
plot_tcn_heatmap(segs, fits, groups, group_colors, sex, normal_sample)
}
\description{
generate a heatmap with total copy number, a NJ tree based on TCN, and a barplot annotating tumor purity
}
