% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_bin_data.R
\name{preprocess_bin_data}
\alias{preprocess_bin_data}
\title{preprocess_bin_data}
\usage{
preprocess_bin_data(
  qdnaseq_data,
  pileup_data,
  phased_bcf,
  sample_map,
  normal_sample,
  sex,
  build,
  max_phaseable_distance,
  min_bin_reads_for_baf,
  blacklisted_regions_file,
  LogR_range_allowed,
  LogR_winsor_percentiles,
  LogR_smooth_bins,
  normal_correction
)
}
\description{
Create preprocessed data from snp-pileup, QDNAseq, and GLIMPSE output.
}
