% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ascat_multipcf.R
\name{run_ascat_multipcf}
\alias{run_ascat_multipcf}
\title{run_ascat_multipcf
Run ascat.asmultipcf to do multi-sample segmentation, returns appended data objected.}
\usage{
run_ascat_multipcf(
  obj,
  penalty = 70,
  seed = as.integer(Sys.time()),
  refine = T,
  selectAlg = "exact",
  overwrite_mpcf = F,
  build = "hg19"
)
}
\description{
run_ascat_multipcf
Run ascat.asmultipcf to do multi-sample segmentation, returns appended data objected.
}
