% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ascat_multipcf.R
\name{run_ascat_multipcf}
\alias{run_ascat_multipcf}
\title{run_ascat_multipcf}
\usage{
run_ascat_multipcf(
  obj,
  build,
  penalty,
  refine,
  selectAlg,
  seed = NA,
  excluded_samples = c()
)
}
\description{
Do multi-sample segmentation using ascat.asmultipcf.
}
