% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_preprocessed_lowpass_data_for_ascat.R
\name{save_preprocessed_lowpass_data_for_ascat}
\alias{save_preprocessed_lowpass_data_for_ascat}
\title{save_preprocessed_lowpass_data_for_ascat}
\usage{
save_preprocessed_lowpass_data_for_ascat(
  preprocessed_data,
  normal_sample,
  sex,
  Tumor_LogR_file,
  Tumor_BAF_file,
  Germline_LogR_file,
  Germline_BAF_file
)
}
\description{
Write preprocessed lowpass data into the format expected for ascat.loadData()
}
